pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "wedley", email = "wedleysilva2105@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
